Move BuildCacheCommandFactory to :build-cache
And its implementation to :core (though it should end up in some build-cache-related subproject eventually).
