+++ subprojects/build-cache/build-cache.gradle.kts
@@ -23,6 +23,7 @@ plugins {
 
 dependencies {
     api(project(":buildCacheBase"))
+    api(project(":snapshots"))
 
     implementation(project(":baseServices"))
     implementation(project(":coreApi"))
