+++ subprojects/core/src/main/java/org/gradle/caching/internal/controller/impl/DefaultBuildCacheCommandFactory.java
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package org.gradle.caching.internal.command.impl;
+package org.gradle.caching.internal.controller.impl;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSortedMap;
 import com.google.common.collect.Interner;
 import org.gradle.caching.BuildCacheKey;
 import org.gradle.caching.internal.CacheableEntity;
-import org.gradle.caching.internal.command.BuildCacheCommandFactory;
+import org.gradle.caching.internal.controller.BuildCacheCommandFactory;
 import org.gradle.caching.internal.controller.BuildCacheLoadCommand;
 import org.gradle.caching.internal.controller.BuildCacheStoreCommand;
 import org.gradle.caching.internal.origin.OriginMetadata;
