+++ subprojects/execution/src/test/groovy/org/gradle/internal/execution/steps/CacheStepTest.groovy
@@ -18,7 +18,7 @@ package org.gradle.internal.execution.steps
 
 import com.google.common.collect.ImmutableList
 import org.gradle.caching.BuildCacheKey
-import org.gradle.caching.internal.command.BuildCacheCommandFactory
+import org.gradle.caching.internal.controller.BuildCacheCommandFactory
 import org.gradle.caching.internal.controller.BuildCacheController
 import org.gradle.caching.internal.controller.BuildCacheLoadCommand
 import org.gradle.caching.internal.controller.BuildCacheStoreCommand
